// src/app/core/api/generated/model/attendanceCalculationPreviewRowDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AttendanceCalculationPreviewRowDto { 
    employeeId?: string;
    employeeName?: string | null;
    employeeCode?: string | null;
    departmentName?: string | null;
    planName?: string | null;
    shiftsSummaryInRange?: string | null;
    calendarDaysInRange?: number;
    daysWithPunches?: number;
    daysPresent?: number;
    totalNetWorkMinutes?: number;
    totalLateMinutes?: number;
    daysAbsentOnScheduledWorkdays?: number;
    daysIncompleteAttendance?: number;
    totalEarlyLeaveMinutes?: number;
    totalOvertimeBeforeMinutes?: number;
    totalOvertimeAfterMinutes?: number;
}