// src/app/core/api/generated/model/commitAttendanceCalculationCommand.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PunchPairingMethod } from './punchPairingMethod';


export interface CommitAttendanceCalculationCommand { 
    from?: string;
    to?: string;
    pairingMethod?: PunchPairingMethod;
    nameFilter?: string | null;
    planId?: string | null;
    shiftId?: string | null;
    note?: string | null;
}